feat(favorites): implement paginated favorites API with enhanced response structure
- Refactored the favorites API to support pagination, allowing users to retrieve their favorite tutors in a paginated format.
- Updated the response structure to include metadata such as total items, current page, page size, total pages, and navigation flags (has_next, has_prev).
- Adjusted the database queries to efficiently handle pagination and eager loading of related tutor profiles and subjects.
- Modified schemas to accommodate the new paginated response format.
- Added tests to ensure the new pagination functionality works as expected and validates input parameters.

This enhancement improves user experience by allowing for better navigation through large sets of favorite tutors.

+++ backend/schemas.py
@@ -524,7 +524,7 @@ class TutorProfileResponse(BaseModel):
     profile_status: str
     rejection_reason: str | None
     approved_at: datetime | None = None
-    average_rating: Decimal
+    average_rating: float
     total_reviews: int
     total_sessions: int
     created_at: datetime
@@ -552,7 +552,7 @@ class TutorPublicProfile(BaseModel):
     bio: str | None
     hourly_rate: Decimal
     experience_years: int
-    average_rating: Decimal
+    average_rating: float
     total_reviews: int
     total_sessions: int
     subjects: list[PublicSubjectItem] = []
